docks(UIKit): placed in a separate folder components that are used only in the presentation && style(all)
