--- my-app/src/components/Icon/ForwarArrow.jsx
@@ -1,15 +1,0 @@
-
-const ForwardArrow = ({width, height, className, name}) => {
-    return (
-        <svg className = {className}
-             name = {name}
-             width={width} 
-             height={height}
-             viewBox="0 0 20 10" 
-             xmlns="http://www.w3.org/2000/svg">
-<path d="M19.7709 4.44705C19.7706 4.44681 19.7704 4.44654 19.7702 4.4463L15.688 0.383808C15.3821 0.0794727 14.8875 0.0806052 14.5831 0.386464C14.2787 0.692284 14.2799 1.18693 14.5857 1.49131L17.3265 4.2188H0.78125C0.349766 4.2188 0 4.56857 0 5.00005C0 5.43154 0.349766 5.7813 0.78125 5.7813H17.3264L14.5857 8.5088C14.2799 8.81317 14.2788 9.30782 14.5831 9.61364C14.8875 9.91954 15.3822 9.92059 15.688 9.6163L19.7702 5.5538C19.7704 5.55357 19.7706 5.55329 19.7709 5.55306C20.0769 5.24767 20.0759 4.75142 19.7709 4.44705Z"/>
-</svg>
-    );
-}
-
-export default ForwardArrow;
