--- my-app/src/components/SideBar/SideBar.module.css
@@ -1,32 +1,0 @@
-.side_bar {
-    width: 160px;
-    height: 100%;
-    background-color: #3d3d3d;
-  }
-
-.side_bar__list {
-  height: auto;
-  margin: 0;
-  padding: 0;
-  list-style: none;
-}
-
-.side_bar__item a{
-  display: block;
-  padding: 10px;
-  text-decoration: none;
-  color: aliceblue;
-  border-bottom: solid 1px #666;
-  transition: 0.5s;
-}
-
-.side_bar__item a:hover {
-  background: #666;
-  padding: 10px 0 10px 20px;
-}
-
-a.active_link {
-  background-color: rgba(128, 125, 125, 0.932);
-  padding: 10px 0 10px 20px;
-}
-
