+++ my-app/src/pages/ButtonDemo/ButtonDemo.jsx
@@ -1,68 +1,67 @@
 import Button from "../../components/Button/Button";
-import s from '../../components/Button/Button.module.css'
-import Icon from "../../components/Icon/Index";
+import Icon from "../../components//UIDemo/Icon/Index";
 
-function ButtonDemo () {
+function ButtonDemo() {
     return (
         <>
             <Button
-                id="my-button" 
+                id="my-button"
                 onClick={() => console.log('log log')}
-                className="someCssCLass" 
+                className="someCssCLass"
                 disable={false}
                 type="primary"
                 title="SAVE"
-                /> 
+            />
 
             <Button
-                id="my-button" 
+                id="my-button"
                 onClick={() => console.log('log log')}
-                className="someCssCLass" 
+                className="someCssCLass"
                 disable={false}
                 type="primary"
                 title="SAVE"
-                endIcon={<Icon name="ForwardArrow" height="14px" width="14px" className={s.btn__icon__end}/>}
-                /> 
+                endIcon={<Icon name="ForwardArrow" height="14px" width="14px"/>}
+            />
 
             <Button
-                id="my-button" 
+                id="my-button"
                 onClick={() => console.log('log log')}
-                className="someCssCLass" 
+                className="someCssCLass"
                 disable={true}
                 type="primary"
                 title="SAVE"
-                startIcon={<Icon name="Lock" height="14px" width="14px" className={s.btn__icon__start}/>}
-                endIcon={<Icon name="ForwardArrow" height="14px" width="14px" className={s.btn__icon__end}/>}
-                /> 
+                startIcon={<Icon name="Lock" height="15px" width="18px" />}
+                endIcon={<Icon name="ForwardArrow" height="14px" width="14px" />}
+            />
 
             <Button
-                id="my-button" 
+                id="my-button"
                 onClick={() => console.log('log log')}
-                className="someCssCLass" 
+                className="someCssCLass"
                 disable={false}
                 type="secondary"
                 title="SAVE"
-                /> 
-            
+            />
+
             <Button
-                id="my-button" 
+                id="my-button"
                 onClick={() => console.log('log log')}
-                className="someCssCLass" 
+                className="someCssCLass"
                 disable={false}
                 type="secondary"
                 title="SAVE"
-                endIcon={<Icon name="ForwardArrow" height="14px" width="14px" className={s.btn__icon__end}/>}
-                /> 
-            
+                endIcon={<Icon name="ForwardArrow" height="14px" width="14px" />}
+            />
+
             <Button
-                id="my-button" 
+                id="my-button"
                 onClick={() => console.log('log log')}
-                className="someCssCLass" 
+                className="someCssCLass"
                 disable={true}
                 type="secondary"
                 title="SAVE"
-                endIcon={<Icon name="ForwardArrow" height="14px" width="14px" className={s.btn__icon__end}/>}
-                /> 
+                endIcon={<Icon name="ForwardArrow" height="14px" width="14px" />}
+            />
         </>
     )
 }
